+++ src/honeychrome/main.py
@@ -136,12 +136,12 @@ def main():
     '''
     define objects for communication between processes
     '''
-    from honeychrome.instrument_driver_components.cytkit_configuration import traces_cache_size, dtype
+    from honeychrome.settings import traces_cache_size, traces_cache_dtype
     from honeychrome.settings import max_events_in_cache, n_channels_per_event
     import honeychrome.settings as settings
 
     # Allocate shared memory block, plus head and tail indices
-    traces_cache_shm = shared_memory.SharedMemory(create=True, size=np.zeros(traces_cache_size, dtype=dtype).nbytes)
+    traces_cache_shm = shared_memory.SharedMemory(create=True, size=np.zeros(traces_cache_size, dtype=traces_cache_dtype).nbytes)
     traces_cache_lock = Lock()
     index_head_traces_cache = mp.Value('i', 0)
     index_tail_traces_cache = mp.Value('i', 0)
